refactor: make TaskStore a trait with a CachedTaskStore impl
