+++ src/main.rs
@@ -8,10 +8,12 @@ use openssl::pkey::{PKey, Private};
 use openssl::x509::X509;
 use persistence::PostgresRepository;
 
+use crate::cached_task_store::CachedTaskStore;
 use crate::state::State;
 use tokio::try_join;
 use tonic::codegen::Arc;
 
+mod cached_task_store;
 mod communicator;
 mod error;
 mod interfaces;
@@ -232,7 +234,9 @@ async fn main() -> Result<(), String> {
         .await
         .expect("Coudln't init postgres repo");
     repo.apply_migrations().expect("Couldn't apply migrations");
-    let state = State::restore(Arc::new(repo))
+    let repo = Arc::new(repo);
+    let task_store = CachedTaskStore::new(repo.clone());
+    let state = State::restore(repo, task_store)
         .await
         .expect("Couldn't initialize State");
     // TODO: remove mutex when DB done
