+++ src/state.rs
@@ -3,6 +3,7 @@ use log::{debug, error, info, warn};
 use std::collections::HashMap;
 use uuid::Uuid;
 
+use crate::cached_task_store::CachedTaskStore;
 use crate::error::Error;
 use crate::persistence::{
     Device, DeviceKind, Group, KeyType, NameValidator, Participant, PersistenceError, ProtocolType,
@@ -21,28 +22,33 @@ use tokio::sync::mpsc::Sender;
 use tonic::codegen::Arc;
 use tonic::Status;
 
-pub struct State {
+pub type State = StateInner<CachedTaskStore>;
+
+pub(crate) struct StateInner<TS: TaskStore> {
     devices: DashMap<Vec<u8>, Device>,
     subscribers: DashMap<Vec<u8>, Sender<Result<proto::Task, Status>>>,
     repo: Arc<dyn Repository + Send + Sync>,
-    task_store: TaskStore,
+    task_store: TS,
     task_last_updates: DashMap<Uuid, u64>,
     device_last_activations: DashMap<Vec<u8>, u64>,
 }
 
-impl State {
-    pub async fn restore(repo: Arc<dyn Repository + Send + Sync>) -> Result<Self, Error> {
+impl<TS: TaskStore + Sync> StateInner<TS> {
+    pub async fn restore(
+        repo: Arc<dyn Repository + Send + Sync>,
+        task_store: TS,
+    ) -> Result<Self, Error> {
         let devices = repo
             .get_devices()
             .await?
             .into_iter()
             .map(|dev| (dev.id.clone(), dev))
             .collect();
-        let state = State {
+        let state = Self {
             devices,
             subscribers: DashMap::new(),
             repo: repo.clone(),
-            task_store: TaskStore::new(repo),
+            task_store,
             task_last_updates: DashMap::new(),
             device_last_activations: DashMap::new(),
         };
@@ -265,8 +271,10 @@ impl State {
     ) -> Result<Vec<proto::Task>, Error> {
         let task_ids = self.repo.get_active_device_tasks(device_id).await?;
         let mut tasks = Vec::new();
-        for task in self.task_store.get_tasks(task_ids).await? {
-            tasks.push(task.await.format(Some(device_id), None));
+        for task_id in task_ids {
+            let task = self.task_store.get_task(&task_id).await?;
+            let task = task.format(Some(device_id), None);
+            tasks.push(task);
         }
         Ok(tasks)
     }
@@ -315,8 +323,10 @@ impl State {
     pub async fn get_formatted_tasks(&self) -> Result<Vec<proto::Task>, Error> {
         let task_ids = self.repo.get_tasks().await?;
         let mut tasks = Vec::new();
-        for task in self.task_store.get_tasks(task_ids).await? {
-            tasks.push(task.await.format(None, None));
+        for task_id in task_ids {
+            let task = self.task_store.get_task(&task_id).await?;
+            let task = task.format(None, None);
+            tasks.push(task);
         }
         Ok(tasks)
     }
@@ -552,7 +562,7 @@ impl State {
 
     pub async fn restart_stale_tasks(&self) -> Result<(), Error> {
         let now = get_timestamp();
-        let task_ids = self
+        let task_ids: Vec<_> = self
             .repo
             .get_restart_candidates()
             .await?
@@ -562,9 +572,8 @@ impl State {
                 now.saturating_sub(last_update) > 30
             })
             .collect();
-        let tasks = self.task_store.get_tasks_mut(task_ids).await?;
-        for task_entry in tasks {
-            let task_entry = &mut *task_entry.await;
+        for task_id in task_ids {
+            let task_entry = &mut *self.task_store.get_task_mut(&task_id).await?;
             let Task::Running(task) = task_entry else {
                 return Err(PersistenceError::DataInconsistencyError(
                     "non-running task in candidates for restart".into(),
